requirements.txt
slack-sdk
slack-bolt
openai==0.27.5
emoji==2.2.0
python-dotenv
cachier
